requirements.txt
aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1676292794306/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1683424013410/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1681349778161/work
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1666764877873/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1674482203741/work
certifi==2023.5.7
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1671179612308/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1685487537961/work
frozenlist @ file:///Users/runner/miniforge3/conda-bld/frozenlist_1667935707665/work
google-api-core @ file:///home/conda/feedstock_root/build_artifacts/google-api-core-split_1670580271439/work
google-api-python-client @ file:///home/conda/feedstock_root/build_artifacts/google-api-python-client_1685447681973/work
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1685078704305/work
google-auth-httplib2 @ file:///home/conda/feedstock_root/build_artifacts/google-auth-httplib2_1654134467987/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1675825735130/work
googleapis-common-protos @ file:///home/conda/feedstock_root/build_artifacts/googleapis-common-protos-feedstock_1672913962683/work
httplib2 @ file:///home/conda/feedstock_root/build_artifacts/httplib2_1679483503307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
multidict @ file:///Users/runner/miniforge3/conda-bld/multidict_1672339475162/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1682210448415/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1666056362788/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1685343038613/work
protobuf==4.23.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1683676063469/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1685514481738/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1680088766131/work
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1658328885051/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1685045523825/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1680081134351/work
uritemplate @ file:///home/conda/feedstock_root/build_artifacts/uritemplate_1634152692041/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
yarl @ file:///Users/runner/miniforge3/conda-bld/yarl_1685191791125/work
